backend/requirements.txt
requests
uvicorn
pydantic
fastapi
python-dotenv